+++ CHANGELOG.md
@@ -13,11 +13,13 @@
 * Fix #5357: adding additional Quantity methods
 * Fix #5635: refined LeaderElector lifecycle and logging
 
+
 #### Dependency Upgrade
 
 #### New Features
 * Fix #5636: Add new extension `open-virtual-networking` to manage resources in `k8s.ovn.org/v1` API group.
 * Fix #5711: Kube API Test - Kubernetes API Server JUnit Test Support
+* Fix #5772: Add openshift model `io.fabric8.openshift.api.model.DeploymentConfigRollback`
 
 #### _**Note**_: Breaking changes
 * KubeSchema and Validation Schema generated classes are no longer annotated with Jackson, Lombok, and Sundrio annotations.